Refine SSE client registration and resume flow
